Disentangle implementation relevant dependencies
