+++ dev/dbstat.h
@@ -2,6 +2,9 @@
 
 #include <string>  //  std::string
 
+#include "column.h"
+#include "table.h"
+
 namespace sqlite_orm {
 #ifdef SQLITE_ENABLE_DBSTAT_VTAB
     struct dbstat {
@@ -16,5 +19,19 @@ namespace sqlite_orm {
         int pgoffset = 0;
         int pgsize = 0;
     };
+
+    inline auto make_dbstat_table() {
+        return make_table("dbstat",
+                          make_column("name", &dbstat::name),
+                          make_column("path", &dbstat::path),
+                          make_column("pageno", &dbstat::pageno),
+                          make_column("pagetype", &dbstat::pagetype),
+                          make_column("ncell", &dbstat::ncell),
+                          make_column("payload", &dbstat::payload),
+                          make_column("unused", &dbstat::unused),
+                          make_column("mx_payload", &dbstat::mx_payload),
+                          make_column("pgoffset", &dbstat::pgoffset),
+                          make_column("pgsize", &dbstat::pgsize));
+    }
 #endif  //  SQLITE_ENABLE_DBSTAT_VTAB
 }
