+++ dev/type_traits.h
@@ -38,6 +38,9 @@ namespace sqlite_orm {
         using type_t = typename T::type;
 
         template<typename T>
+        using field_type_t = typename T::field_type;
+
+        template<typename T>
         using object_type_t = typename T::object_type;
 
         template<typename T>
