+++ not_single_header_include/sqlite_orm/sqlite_orm.h
@@ -1,24 +1,13 @@
 #pragma once
 
 #include "../../dev/start_macros.h"
+#include "../../dev/type_traits.h"
 #include "../../dev/error_code.h"
-#include "../../dev/ast/upsert_clause.h"
-#include "../../dev/ast/excluded.h"
 #include "../../dev/tuple_helper/tuple_helper.h"
 #include "../../dev/tuple_helper/find_in_tuple.h"
 #include "../../dev/tuple_helper/tuple_transformer.h"
 #include "../../dev/tuple_helper/count_tuple.h"
 #include "../../dev/tuple_helper/same_or_void.h"
-#include "../../dev/member_traits/is_field_member_pointer.h"
-#include "../../dev/member_traits/field_member_traits.h"
-#include "../../dev/member_traits/getters.h"
-#include "../../dev/member_traits/setters.h"
-#include "../../dev/member_traits/is_getter.h"
-#include "../../dev/member_traits/is_setter.h"
-#include "../../dev/member_traits/getter_traits.h"
-#include "../../dev/member_traits/setter_traits.h"
-#include "../../dev/member_traits/member_traits.h"
-#include "../../dev/xdestroy_handling.h"
 #include "../../dev/type_printer.h"
 #include "../../dev/collate_argument.h"
 #include "../../dev/constraints.h"
@@ -34,6 +23,7 @@
 #include "../../dev/typed_comparator.h"
 #include "../../dev/select_constraints.h"
 #include "../../dev/table_info.h"
+#include "../../dev/triggers.h"
 #include "../../dev/statement_finalizer.h"
 #include "../../dev/arithmetic_tag.h"
 #include "../../dev/pointer_value.h"
@@ -52,3 +42,4 @@
 #include "../../dev/node_tuple.h"
 #include "../../dev/get_prepared_statement.h"
 #include "../../dev/carray.h"
+#include "../../dev/interface_definitions.h"
